# [24.0 backport] pkg/fileutils: GetTotalUsedFds: reduce allocations