@neersighted 2023-06-30
I just ran into the containerd log changes myself in https://github.com/moby/moby/pull/45857; personally I think it's more idiomatic to fix the conflict in the cherry-picked commit itself, rather than synthesize a single-file cherry-pick to make a downstream cherry-pick 'clean.'